Extract literals.
